backend/requirements.txt
# ─── Backend Dependencies ────────────────────────────────────────────
# Blockchain Voting System with Face Verification
# Install with: pip install -r requirements.txt

fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.36
web3==7.6.0
deepface==0.0.93
opencv-python-headless==4.10.0.84
Pillow==11.1.0
numpy==1.26.4
